Add debugging and other methods for telegram

+++ telegram.py
@@ -15,31 +15,32 @@ teleBot = Blueprint('teleBot', __name__, template_folder='templates')
 
 @teleBot.route('/assassins/telegram/update/', methods = ["POST"])
 def telegramUpdate():
-    data = request.get_json()
-    chatID = data["message"]["chat"]["id"]
-    print(data)
     conn = psycopg2.connect(connStr)
     conn.autocommit = True
     cur = conn.cursor()
 
-    if chatID not in ids:
-        ids.add(chatID)
-    if "text" in data["message"] and data["message"]["text"][0:7] == "/status":
-        # Send update
-        print("Received status update request")
-        # Fetch status
-        outputStr = fetchStatus(cur)
-        sendMsg(chatID, outputStr)
-    elif "text" in data["message"] and data["message"]["text"][0:6] == "/start":
-        user_hash = data["message"]["text"][7:]
-        #retrieve associated user_id, store tele_chat_id
-        
-        cur.execute("SELECT user_id FROM tele_ids WHERE tele_hash = %s", (user_hash,))
-        data = cur.fetchone()
-        if data is not None:
-            user_id = data[0]
-            cur.execute("UPDATE users SET user_telegram = %s WHERE user_id = %s", (chatID, user_id,))
-            cur.execute("DELETE FROM tele_ids WHERE user_id = %s", (user_id,))
+    data = request.get_json()
+    print(data)
+    if "message" in data:
+        chatID = data["message"]["chat"]["id"]
+        if chatID not in ids:
+            ids.add(chatID)
+        if "text" in data["message"] and data["message"]["text"][0:7] == "/status":
+            # Send update
+            print("Received status update request")
+            # Fetch status
+            outputStr = fetchStatus(cur)
+            sendMsg(chatID, outputStr)
+        elif "text" in data["message"] and data["message"]["text"][0:6] == "/start":
+            user_hash = data["message"]["text"][7:]
+            #retrieve associated user_id, store tele_chat_id
+            
+            cur.execute("SELECT user_id FROM tele_ids WHERE tele_hash = %s", (user_hash,))
+            data = cur.fetchone()
+            if data is not None:
+                user_id = data[0]
+                cur.execute("UPDATE users SET user_telegram = %s WHERE user_id = %s", (chatID, user_id,))
+                cur.execute("DELETE FROM tele_ids WHERE user_id = %s", (user_id,))
     cur.close()
 
     return json.dumps({'success':True}), 200, {'ContentType':'application/json'} 
@@ -48,6 +49,38 @@ def sendMsg(id, msg):
     r = requests.post(
         telegramBotURL+"sendMessage", 
         data = {"chat_id": id, "text": msg, "parse_mode" : "Markdown"})
+    print(r.text)
+
+@teleBot.route("/testing/")
+def testingMethod():
+    sendKillConfirmMsg(167223959, "Hello testing")
+    return "{}"
+
+def sendKillConfirmMsg(id, msg):
+    keyboard = {
+        "inline_keyboard": [
+            [{
+                "text": "Accept", 
+                "callback_data": json.dumps({
+                    "decision": "accept", 
+                    "id": 3
+                })
+            }], [{
+                "text": "Reject", 
+                "callback_data": json.dumps({
+                    "decision": "reject", 
+                    "id": 3
+                })
+            }]
+        ]
+    }
+
+    r = requests.post(telegramBotURL+"sendMessage", data = {
+        "chat_id": id, 
+        "text": msg, 
+        "parse_mode": "Markdown", 
+        "reply_markup": json.dumps(keyboard)
+    })
     print(r.text)
 
 def fetchStatus(cur):
